Move Internal out of Git.Types
