+++ src/Git/Refs.hs
@@ -15,13 +15,13 @@ import System.Directory (listDirectory)
 import System.IO (withBinaryFile, IOMode(..))
 import System.IO.Error (isDoesNotExistError)
 
+import Git.Internal ()
 import Git.Objects (Commit)
 -- FIXME: sha1HexParser should probably live nearer Git.Sha1...
 import Git.Objects.Serialise (sha1HexParser)
 import Git.Repository (Repo, repoRefsPath, repoHeadPath)
 import Git.Sha1 (Sha1)
 import qualified Git.Sha1 as Sha1
-import Git.Types.Internal ()
 
 
 class Sha1Pointer a where
